pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "secure-files-cli"
version = "1.0.0"
description = "Advanced file encryption tool with AES-256-GCM, PBKDF2, and comprehensive security features"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Eternal0404", email = "eternal0404@example.com"}
]
keywords = ["encryption", "security", "aes", "gcm", "pbkdf2", "cli", "file-encryption"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
    "Topic :: System :: Archiving :: Backup",
    "Topic :: Utilities",
]
requires-python = ">=3.8"
dependencies = [
    "pycryptodome>=3.19.0",
    "toml>=0.10.0",
    "pyyaml>=6.0",
    "tqdm>=4.64.0",
    "rich>=13.0.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/Eternal0404/secure-files-cli"
Repository = "https://github.com/Eternal0404/secure-files-cli"
Documentation = "https://github.com/Eternal0404/secure-files-cli/blob/main/README.md"
"Bug Tracker" = "https://github.com/Eternal0404/secure-files-cli/issues"

[project.scripts]
secure-files-cli = "secure_files_cli.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["secure_files_cli*"]

[tool.setuptools.package-data]
secure_files_cli = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=secure_files_cli --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["secure_files_cli"]
omit = [
    "*/tests/*",
    "*/test_*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]